# docs-site/openapi.yaml
openapi: 3.0.3
info:
  title: LogPulse API
  version: 1.1.0
  description: |
    REST API for LogPulse — a centralized log monitoring platform.

    Three authentication realms exist:

    | Realm | Mechanism | Used by |
    |-------|-----------|---------|
    | Dashboard JWT | `Authorization: Bearer <jwt>` (24h expiry, HS256) | everything under `/api/**` except ingest |
    | Ingest API key | `Authorization: Bearer <application API key>` | `POST /api/logs` only |
    | OAuth redirect | browser 302 flow | `/api/auth/{google,github}/*` |

    There is no public signup: an admin invites users, and the invited person
    activates the account via `POST /api/auth/accept-invite` or by signing in
    with Google/GitHub.

    Errors are always `{"error": "<human readable message>"}`.
servers:
  - url: http://localhost:8080
    description: Local development
  # INFERRED: production URL depends on your deployment; set your own.
  - url: https://your-logpulse-host
    description: Production (replace with your host)
...
paths:
...
  /api/incidents:
    get:
      tags: [Incidents]
      summary: List incidents
      description: Newest first, max 200. Without a status filter, all statuses are returned.
      operationId: listIncidents
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, open, acknowledged, resolved] }
          description: '`active` = open + acknowledged.'
      responses:
        '200':
          description: Incident list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Incident' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Incidents]
      summary: Declare an incident manually
      operationId: createIncident
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                severity: { type: string, enum: [critical, error, warning, info], default: error }
                affected_apps:
                  type: string
                  description: JSON-encoded array of app names, e.g. '["api","worker"]'
                summary: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/incidents/{id}:
    get:
      tags: [Incidents]
      summary: Get one incident
      operationId: getIncident
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IncidentId' }]
      responses:
        '200':
          description: The incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Incidents]
      summary: Update status / severity / fields
      description: >
        Setting status to `resolved` stamps `resolved_at`; moving back to
        open/acknowledged clears it.
      operationId: updateIncident
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IncidentId' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [open, acknowledged, resolved] }
                severity: { type: string, enum: [critical, error, warning, info] }
                title: { type: string }
                summary: { type: string }
      responses:
        '200':
          description: Updated incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/alerts/parse:
    post:
      tags: [Alerts]
      summary: Parse plain language into a draft rule (AI-8)
      description: >
        Deterministic parser — no LLM, no data leaves the deployment. Returns
        a draft for human review; nothing is activated until you POST it to
        `/api/alerts`.
      operationId: parseAlert
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: Alert me when payments-api has more than 5 errors mentioning timeout within 2 minutes
      responses:
        '200':
          description: Draft rule plus the matched application id (or null)
          content:
            application/json:
              schema:
                type: object
                properties:
                  draft: { $ref: '#/components/schemas/ParsedAlertDraft' }
                  application_id:
                    type: integer
                    nullable: true
        '400': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard JWT from /api/auth/login — 24h expiry (HS256).
    ingestApiKey:
      type: http
      scheme: bearer
      description: Per-application API key from POST /api/applications.
  parameters:
    IncidentId: { name: id, in: path, required: true, schema: { type: integer } }
...
  responses:
    ValidationError:
      description: Request body failed validation (message is the validator error)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Missing/expired JWT ("missing token" / "invalid or expired token")
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotFound:
      description: Resource does not exist
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Incident:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        severity: { type: string, enum: [critical, error, warning, info] }
        status: { type: string, enum: [open, acknowledged, resolved] }
        source: { type: string, enum: [anomaly, correlation, alert, manual] }
        summary: { type: string }
        evidence: { type: string, description: JSON-encoded array (reserved for stored evidence) }
        affected_apps: { type: string, description: JSON-encoded array of app names }
        alert_event_ids: { type: string, description: JSON-encoded array of linked AlertEvent ids (AI-6) }
        started_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
        created_by_user_id: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ParsedAlertDraft:
      type: object
      properties:
        application_name: { type: string, description: Empty when no registered app matched }
        level: { type: string, enum: [critical, error, warning, info, debug] }
        keyword: { type: string }
        threshold: { type: integer }
        window_seconds: { type: integer }
        cooldown_seconds: { type: integer }
        confidence: { type: number, format: float, description: 0–1 heuristic }
        notes: { type: array, items: { type: string }, description: What the parser wants a human to review }
...